refactor: remove legacy collection
- and remove all dependent codes and relevant tests
- update the script that generates the filters.yaml
- see RHINENG-6982
