+++ insights/collect.py
@@ -525,7 +525,7 @@ def collect(client_config=None, rm_conf=None, tmp_path=None,
 
     collect_errors = _parse_broker_exceptions(broker, EXCEPTIONS_TO_REPORT)
 
-    cleaner.generate_report(archive_name, client_config.rhsm_facts_file) if cleaner else None
+    cleaner.generate_report(archive_name) if cleaner else None
 
     if compress:
         return create_archive(output_path), collect_errors
