+++ insights/tests/client/core_collector/test_write_metadata.py
@@ -2,13 +2,13 @@ import six
 import mock
 from insights.client.constants import InsightsConstants as constants
 from insights.client.config import InsightsConfig
-from insights.client.data_collector import DataCollector
+from insights.client.core_collector import CoreCollector
 from mock.mock import call
 from mock.mock import patch
 
 
-@patch('insights.client.data_collector.os.remove')
-@patch('insights.client.data_collector.InsightsArchive')
+@patch('insights.client.core_collector.os.remove')
+@patch('insights.client.core_collector.InsightsArchive')
 def test_egg_release_file_read_and_written(archive, remove):
     '''
     Verify the egg release file is read from file and
@@ -22,14 +22,14 @@ def test_egg_release_file_read_and_written(archive, remove):
     with patch(open_name, create=True) as mock_open:
         mock_open.side_effect = [mock.mock_open(read_data='/testvalue').return_value]
         c = InsightsConfig()
-        d = DataCollector(c)
+        d = CoreCollector(c)
         d._write_egg_release()
         remove.assert_called_once_with(constants.egg_release_file)
         d.archive.add_metadata_to_archive.assert_called_once_with('/testvalue', '/egg_release')
 
 
-@patch('insights.client.data_collector.os.remove')
-@patch('insights.client.data_collector.InsightsArchive')
+@patch('insights.client.core_collector.os.remove')
+@patch('insights.client.core_collector.InsightsArchive')
 def test_egg_release_file_read_and_written_no_delete(archive, remove):
     '''
     Verify the egg release file is read from file and
@@ -45,14 +45,14 @@ def test_egg_release_file_read_and_written_no_delete(archive, remove):
     with patch(open_name, create=True) as mock_open:
         mock_open.side_effect = [mock.mock_open(read_data='/testvalue').return_value]
         c = InsightsConfig()
-        d = DataCollector(c)
+        d = CoreCollector(c)
         d._write_egg_release()
         remove.assert_called_once_with(constants.egg_release_file)
         d.archive.add_metadata_to_archive.assert_called_once_with('/testvalue', '/egg_release')
 
 
-@patch('insights.client.data_collector.os.remove')
-@patch('insights.client.data_collector.InsightsArchive')
+@patch('insights.client.core_collector.os.remove')
+@patch('insights.client.core_collector.InsightsArchive')
 def test_egg_release_file_read_and_written_no_read(archive, remove):
     '''
     Verify that when the egg release file cannot be read,
@@ -68,14 +68,14 @@ def test_egg_release_file_read_and_written_no_read(archive, remove):
     with patch(open_name, create=True) as mock_open:
         mock_open.side_effect = IOError('test')
         c = InsightsConfig()
-        d = DataCollector(c)
+        d = CoreCollector(c)
         d._write_egg_release()
         remove.assert_called_once_with(constants.egg_release_file)
         d.archive.add_metadata_to_archive.assert_called_once_with('', '/egg_release')
 
 
-@patch('insights.client.data_collector.os.remove')
-@patch('insights.client.data_collector.InsightsArchive')
+@patch('insights.client.core_collector.os.remove')
+@patch('insights.client.core_collector.InsightsArchive')
 def test_egg_release_file_read_memory_error(archive, remove):
     '''
     Verify that a memory error on the egg release file read is not
@@ -91,15 +91,15 @@ def test_egg_release_file_read_memory_error(archive, remove):
         file_mock.read.side_effect = MemoryError()
         mock_open.side_effect = [file_mock]
         c = InsightsConfig()
-        d = DataCollector(c)
+        d = CoreCollector(c)
         d._write_egg_release()
         remove.assert_called_once_with(constants.egg_release_file)
         d.archive.add_metadata_to_archive.assert_called_once_with('', '/egg_release')
 
 
-@patch('insights.client.data_collector.os.remove')
+@patch('insights.client.core_collector.os.remove')
 @patch(
-    'insights.client.data_collector.InsightsArchive',
+    'insights.client.core_collector.InsightsArchive',
     **{'return_value.add_metadata_to_archive.side_effect': [OSError('[Errno 28] No space left on device'), None]})
 def test_egg_release_file_write_os_error(archive, remove):
     '''
@@ -114,7 +114,7 @@ def test_egg_release_file_write_os_error(archive, remove):
     with patch(open_name, create=True) as mock_open:
         mock_open.side_effect = [mock.mock_open(read_data='/testvalue').return_value]
         c = InsightsConfig()
-        d = DataCollector(c)
+        d = CoreCollector(c)
         d._write_egg_release()
         remove.assert_called_once_with(constants.egg_release_file)
         failed_call = call('/testvalue', '/egg_release')
