+++ insights/tests/client/phase/test_post_update.py
@@ -24,8 +24,7 @@ def patch_insights_config(old_function):
                        "return_value.load_all.return_value.list_specs": False,
                        "return_value.load_all.return_value.show_results": False,
                        "return_value.load_all.return_value.check_results": False,
-                       "return_value.load_all.return_value.no_upload": False,
-                       "return_value.load_all.return_value.core_collect": False})
+                       "return_value.load_all.return_value.no_upload": False})
     return patcher(old_function)
 
 # DRY this at some point... for the love of god
