+++ insights/tests/client/phase/test_update.py
@@ -12,13 +12,11 @@ def test_update_payload_on(insights_config, insights_client):
     Rules are not updated when a payload is uploaded
     """
     insights_config.return_value.load_all.return_value.payload = True
-    insights_config.return_value.load_all.return_value.core_collect = False
     try:
         update()
     except SystemExit:
         pass
     insights_client.return_value.update.assert_called_once()
-    insights_client.return_value.update_rules.assert_not_called()
 
 
 @patch("insights.client.phase.v1.InsightsClient")
@@ -28,45 +26,11 @@ def test_update_payload_off(insights_config, insights_client):
     Rules are updated in normal operation (no payload)
     """
     insights_config.return_value.load_all.return_value.payload = False
-    insights_config.return_value.load_all.return_value.core_collect = False
     try:
         update()
     except SystemExit:
         pass
     insights_client.return_value.update.assert_called_once()
-    insights_client.return_value.update_rules.assert_called_once()
-
-
-@patch("insights.client.phase.v1.InsightsClient")
-@patch("insights.client.phase.v1.InsightsConfig")
-def test_update_core_collect_on(insights_config, insights_client):
-    """
-    Rules ARE updated when using core collection
-    """
-    insights_config.return_value.load_all.return_value.payload = False
-    insights_config.return_value.load_all.return_value.core_collect = True
-    try:
-        update()
-    except SystemExit:
-        pass
-    insights_client.return_value.update.assert_called_once()
-    insights_client.return_value.update_rules.assert_not_called()
-
-
-@patch("insights.client.phase.v1.InsightsClient")
-@patch("insights.client.phase.v1.InsightsConfig")
-def test_update_core_collect_off(insights_config, insights_client):
-    """
-    Rules are updated when using classic collection
-    """
-    insights_config.return_value.load_all.return_value.payload = False
-    insights_config.return_value.load_all.return_value.core_collect = False
-    try:
-        update()
-    except SystemExit:
-        pass
-    insights_client.return_value.update.assert_called_once()
-    insights_client.return_value.update_rules.assert_called_once()
 
 
 @patch("insights.client.phase.v1.InsightsClient")
