+++ insights/tests/core/spec_cleaner/test_clean_content.py
@@ -8,9 +8,8 @@ test_data_sensitive = 'test \n\n\nabcd\n1234\npassword: p4ssw0rd\n'.splitlines()
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_line_changed_password(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_line_changed_password(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     pp = Cleaner(conf, {})
     ret = pp.clean_content(test_data_sensitive, [])
@@ -21,9 +20,8 @@ def test_redact_line_changed_password(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_keyword_empty_not_change(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_keyword_empty_not_change(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     pp = Cleaner(conf, {})  # empty keywords
     ret = pp.clean_content(test_data, [])
@@ -32,9 +30,8 @@ def test_redact_keyword_empty_not_change(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_keyword_changed_keyword(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_keyword_changed_keyword(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     pp = Cleaner(conf, {'keywords': ['test']})
     ret = pp.clean_content(test_data, [])
@@ -46,9 +43,8 @@ def test_redact_keyword_changed_keyword(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_keyword_no_such_keyword_to_change(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_keyword_no_such_keyword_to_change(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     pp = Cleaner(conf, {'keywords': ['t_e_s_t']})  # no such keyword
     ret = pp.clean_content(test_data, [])
@@ -57,9 +53,8 @@ def test_redact_keyword_no_such_keyword_to_change(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_keyword_disabled_by_no_redact(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_keyword_disabled_by_no_redact(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     pp = Cleaner(conf, {'keywords': 'test'})
     ret = pp.clean_content(test_data, [], no_redact=True)
@@ -68,9 +63,8 @@ def test_redact_keyword_disabled_by_no_redact(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_patterns_exclude_regex(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_patterns_exclude_regex(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     rm_conf = {'patterns': {'regex': ['12.*4', '^abcd']}}
     pp = Cleaner(conf, rm_conf)
@@ -82,9 +76,8 @@ def test_redact_patterns_exclude_regex(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_result_empty(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_result_empty(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     rm_conf = {'patterns': {'regex': [' ', '12.*4', '^abcd', 'test', 'pwd', 'w0rd']}}
     pp = Cleaner(conf, rm_conf)
@@ -94,9 +87,8 @@ def test_redact_result_empty(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_patterns_exclude_no_regex(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_patterns_exclude_no_regex(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     rm_conf = {'patterns': ['1234', 'abcde']}
     pp = Cleaner(conf, rm_conf)
@@ -108,9 +100,8 @@ def test_redact_patterns_exclude_no_regex(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_patterns_exclude_empty(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_patterns_exclude_empty(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     rm_conf = {'patterns': []}
     pp = Cleaner(conf, rm_conf)
@@ -120,9 +111,8 @@ def test_redact_patterns_exclude_empty(core_collect, obfuscate):
 
 
 @mark.parametrize("obfuscate", [True, False])
-@mark.parametrize("core_collect", [True, False])
-def test_redact_exclude_none(core_collect, obfuscate):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate)
+def test_redact_exclude_none(obfuscate):
+    conf = InsightsConfig(obfuscate=obfuscate)
 
     pp = Cleaner(conf, None)
     ret = pp.clean_content(test_data, [])
