+++ insights/tests/core/spec_cleaner/test_reports.py
@@ -7,6 +7,7 @@ from pytest import mark
 
 from insights.client.archive import InsightsArchive
 from insights.client.config import InsightsConfig
+from insights.client.constants import InsightsConstants as constants
 from insights.core.spec_cleaner import Cleaner
 
 hostname = "report.test.com"
@@ -19,11 +20,10 @@ test_file_data = 'ip: 10.0.2.155\ntestword\n{0}'.format(hostname)
         (False, False), (True, False), (True, True),
     ]
 )
-@mark.parametrize("core_collect", [True, False])
 @mark.parametrize("test_umask", [0o000, 0o022])
-def test_rhsm_facts(test_umask, core_collect, obfuscate, obfuscate_hostname):
-    rhsm_facts_file = '/tmp/insights_test_rhsm.facts'
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate,
+def test_rhsm_facts(test_umask, obfuscate, obfuscate_hostname):
+    rhsm_facts_file = constants.rhsm_facts_file = '/tmp/insights_test_rhsm.facts'
+    conf = InsightsConfig(obfuscate=obfuscate,
                           obfuscate_hostname=obfuscate_hostname)
     arch = InsightsArchive(conf)
     arch.create_archive_dir()
@@ -36,7 +36,7 @@ def test_rhsm_facts(test_umask, core_collect, obfuscate, obfuscate_hostname):
     old_umask = os.umask(test_umask)
     pp = Cleaner(conf, {'keywords': ['testword']}, hostname)
     pp.clean_file(test_file, [])
-    pp.generate_report(arch.archive_name, rhsm_facts_file)
+    pp.generate_report(arch.archive_name)
     arch.delete_archive_dir()
 
     umask_after_test = os.umask(old_umask)
@@ -69,7 +69,7 @@ def test_rhsm_facts(test_umask, core_collect, obfuscate, obfuscate_hostname):
         assert kws[0]['original'] == 'testword'
         assert kws[0]['obfuscated'] == 'keyword0'
 
-    os.unlink(rhsm_facts_file)
+    os.unlink(constants.rhsm_facts_file)
 
 
 @mark.parametrize("rm_conf", [{}, {'keywords': ['testword']}])
@@ -79,10 +79,9 @@ def test_rhsm_facts(test_umask, core_collect, obfuscate, obfuscate_hostname):
         (False, False), (True, False), (True, True),
     ]
 )
-@mark.parametrize("core_collect", [True, False])
 @patch('insights.core.spec_cleaner.Cleaner.generate_rhsm_facts', return_value=None)
-def test_all_csv_reports(rhsm_facts, core_collect, obfuscate, rm_conf, obfuscate_hostname):
-    conf = InsightsConfig(core_collect=core_collect, obfuscate=obfuscate,
+def test_all_csv_reports(rhsm_facts, obfuscate, rm_conf, obfuscate_hostname):
+    conf = InsightsConfig(obfuscate=obfuscate,
                           obfuscate_hostname=obfuscate_hostname)
     arch = InsightsArchive(conf)
     arch.create_archive_dir()
@@ -104,7 +103,7 @@ def test_all_csv_reports(rhsm_facts, core_collect, obfuscate, rm_conf, obfuscate
         os.unlink(kw_report_file)
 
     pp.clean_file(test_file, [])
-    pp.generate_report(arch.archive_name, '/dev/null')
+    pp.generate_report(arch.archive_name)
     arch.delete_archive_dir()
 
     if obfuscate:
